fix clippy: drop unnecessary u32→u32 cast and 1*1*3 identity-op
Three clippy errors surfaced after the round-3 revert:
* src/ans_encoder.rs:475 — `(state & 0xFFFF) as u32` is a no-op cast.
* src/ans_encoder.rs:847 — `(i % 8) as u16` for `i: u16` is a no-op.
* src/encoder.rs:892 — `vec![128u8; 1 * 1 * 3]` triggers identity_op.

All three are stylistic cleanups, no behaviour change.

+++ src/ans_encoder.rs
@@ -472,7 +472,7 @@ pub fn encode_symbols_with_extras(
 
         let max_state: u64 = (p as u64) << 20;
         while (state as u64) >= max_state {
-            stack.push(((state & 0xFFFF) as u32, 16));
+            stack.push((state & 0xFFFF, 16));
             state >>= ANS_REFILL_BITS;
         }
         state = inverse_update(state, s, p, inv_alias, alias)?;
@@ -844,7 +844,7 @@ mod tests {
         let alias = AliasTable::build(&d, log_alpha).unwrap();
         let inv = build_inverse_alias(&d, &alias).unwrap();
 
-        let symbols: Vec<u16> = (0..50u16).map(|i| (i % 8) as u16).collect();
+        let symbols: Vec<u16> = (0..50u16).map(|i| i % 8).collect();
         let mut bw = BitWriter::new();
         encode_symbols(&mut bw, &symbols, &d, &inv, &alias).unwrap();
         let bytes = bw.finish();
